+++ package.json
@@ -32,7 +32,7 @@
     "webpack": "3.6.0"
   },
   "dependencies": {
-    "browser-core": "https://github.com/remusao/broxy/releases/download/0.0.1/browser-core-1.0.8.tgz",
+    "browser-core": "https://github.com/remusao/broxy/releases/download/0.0.1/electron-webrequest-0.1.0.tgz",
     "electron": "1.7.8",
     "fast-url-parser": "1.1.3",
     "node-fetch": "1.7.3",
